Simplify blockfrost setup
