+++ hydra-node/src/Hydra/Chain/Direct.hs
@@ -83,12 +83,12 @@ import Hydra.Chain.Direct.TimeHandle (queryTimeHandle)
 import Hydra.Chain.Direct.Util (
   readKeyPair,
  )
-import Hydra.Chain.Direct.Wallet (newTinyWallet)
-import Hydra.Chain.ScriptRegistry (queryScriptRegistry)
-import Hydra.Chain.Wallet (
+import Hydra.Chain.Direct.Wallet (
   TinyWallet (..),
   WalletInfoOnChain (..),
+  newTinyWallet,
  )
+import Hydra.Chain.ScriptRegistry (queryScriptRegistry)
 import Hydra.Logging (Tracer, traceWith)
 import Hydra.Node.Util (readKeyPair)
 import Hydra.Options (DirectChainConfig (..))
