+++ hydra-node/test/Hydra/Chain/Direct/WalletSpec.hs
@@ -41,21 +41,18 @@ import Hydra.Cardano.Api qualified as Api
 import Hydra.Cardano.Api.Prelude (fromShelleyPaymentCredential)
 import Hydra.Cardano.Api.Pretty (renderTx)
 import Hydra.Cardano.Api.Tx (signTx, toLedgerTx)
-import Hydra.Chain.Blockfrost.Wallet qualified as Blockfrost
 import Hydra.Chain.CardanoClient (QueryPoint (..))
 import Hydra.Chain.Direct.Wallet (
-  applyTxs,
-  coverFee_,
-  findLargestUTxO,
-  newTinyWallet,
- )
-import Hydra.Chain.Wallet (
   Address,
   ChainQuery,
   TinyWallet (..),
   TxIn,
   TxOut,
   WalletInfoOnChain (..),
+  applyTxs,
+  coverFee_,
+  findLargestUTxO,
+  newTinyWallet,
  )
 import Test.Hydra.Tx.Fixture qualified as Fixture
 import Test.Hydra.Tx.Gen (genKeyPair, genOneUTxOFor, genSigningKey)
